refactor(providers): extract fallback routing module (#188)
